requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-celery-beat==2.4.0
django-colorfield==0.8.0
-e git+https://github.com/jazzband/django-debug-toolbar.git@36a0a75e668248be9588779360823f8f283851ba#egg=django_debug_toolbar
django-filter==22.1
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
openapi-codec==1.3.2
Pillow==9.4.0
prompt-toolkit==3.0.38
PyJWT==2.6.0
pyotp==2.8.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
